+++ integration-tests/src/test/java/arez/integration/sting/package_access/other/MyDependency.java
@@ -1,0 +1,11 @@
+package arez.integration.sting.package_access.other;
+
+import sting.Injectable;
+
+@Injectable
+class MyDependency
+{
+  MyDependency()
+  {
+  }
+}