+++ integration-tests/src/test/java/arez/integration/sting/package_access/other/TestComponent.java
@@ -1,0 +1,25 @@
+
+package arez.integration.sting.package_access.other;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.Feature;
+import arez.annotations.Observable;
+import arez.annotations.PostConstruct;
+
+@ArezComponent( sting = Feature.ENABLE )
+public abstract class TestComponent
+{
+  TestComponent( final MyDependency myDependency )
+  {
+  }
+
+  @PostConstruct
+  final void postConstruct()
+  {
+  }
+
+  @Observable
+  abstract String getValue();
+
+  abstract void setValue( String value );
+}